pyproject.toml
[tool.poetry]
name = "idk"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
KahootPY = "^0.1.4"
git-kahoot = "^0.2.3"
pygame = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
